ESP32: stop using event descriptors
Event descriptors are a legacy concept, so remove them. Some ugly
workarounds are needed now in socket_driver.
Also remove socket code from libAtomVM so ESP32 sockets can be
refactored independently from unix sockets.
