+++ src/libAtomVM/CMakeLists.txt
@@ -31,8 +31,6 @@ if(${CMAKE_GENERATOR} STREQUAL "Xcode")
         platform_nifs.h
         port.h
         scheduler.h
-        socket.h
-        socket_driver.h
         sys.h
         term_typedef.h
         term.h
@@ -62,7 +60,6 @@ set(SOURCE_FILES
     nifs.c
     port.c
     scheduler.c
-    socket.c
     term.c
     timer_wheel.c
     valueshashtable.c
